+++ packages/react-client/README.md
@@ -1,4 +1,4 @@
-# react-flight
+# react-client
 
 This is an experimental package for consuming custom React streaming models.
