+++ packages/react-client/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "react-flight",
+  "name": "react-client",
   "description": "React package for consuming streaming models.",
   "version": "0.1.0",
   "private": true,
@@ -12,14 +12,13 @@
   "files": [
     "LICENSE",
     "README.md",
-    "index.js",
+    "flight.js",
     "cjs/"
   ],
-  "main": "index.js",
   "repository": {
     "type" : "git",
     "url" : "https://github.com/facebook/react.git",
-    "directory": "packages/react-flight"
+    "directory": "packages/react-client"
   },
   "engines": {
     "node": ">=0.10.0"
