+++ packages/react-client/src/ReactFlightClientHostConfigBrowser.js
