--- packages/react-flight/npm/index.js
@@ -1,7 +1,0 @@
-'use strict';
-
-if (process.env.NODE_ENV === 'production') {
-  module.exports = require('./cjs/react-flight.production.min.js');
-} else {
-  module.exports = require('./cjs/react-flight.development.js');
-}
